Random Sprite Script
Script attached to the "crystal" objects, the collectables in the game. The script makes the game choose between a specified amount of sprites to use for the object. The crystal object has 5 sprites set up in Unity, each a different colour, so each time the game starts the colour of the crystals change to a random colour.